requirements.txt
Django>=4.2.5
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
drf-nested-routers>=0.1.0
psycopg2>=2.9.9
drf-spectacular>=0.26
django-storages>=1.14.2
requests>=2.31.0
pydantic==1.10.12